requirements.txt
pymediainfo
elasticsearch-dsl >=7.0,<8.0
PyYAML
